requirements.txt
pillow
pandas
fastapi
matplotlib
seaborn
numpy
colorama
tqdm
torch
torchvision
torchaudio
anthropic